8. Complex types - functions

+++ 0x00-python_variable_annotations/8-make_multiplier.py
@@ -3,7 +3,7 @@
 from typing import Callable
 
 
-def make_multiplier(multiplier: int) -> Callable[[float], float]:
+def make_multiplier(multiplier: float) -> Callable[[float], float]:
     """
     A function  that takes a float multiplier as argument
     and returns a function that multiplies a float by multiplier.
